refactor(questionnaires): split builder page state orchestration
